requirements.txt
Django==1.6.7
South==1.0
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-admin-bootstrapped==2.0.5
django-debug-toolbar== 1.2.1
django-registration==1.0
django-sslify==0.2.3
docutils==0.11
gunicorn==19.1.1
mock==1.0.1
psycopg2==2.5.4

# Transitive dependencies
sqlparse==0.1.13
static3==0.5.1

